pyproject.toml
[tool.poetry]
name = "genai-1"
version = "0.1.0"
description = ""
authors = ["Daniel Aguilera"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.5"
neo4j = "^5.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
